requirements.txt
selenium
openai
pytest
python-dotenv